Move Trigger into a new file
